+++ CHANGELOG.md
@@ -5,5 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Chaned
+- Require arrow parens '[as needed](https://github.com/airbnb/javascript#arrows--one-arg-parens)'
+
 ### Added
 - Created td-eslint-config